feat(F-178):  reorganize imports in review router and service for clarity
